media: rockchip: rga: Add media controller support, as a scaler
This adds media controller support for the Rockchip RGA driver, using the
dedicated M2M helpers. This is useful for userspace to know which kind of
M2M device the RGA is.

It is described as a scaling unit, although the hardware is also a pixel
encoding converter as well as a pixel formatter, but scaling is probably
the main function to highlight (and the more interesting to offload).

Other media entity nodes may be added later to describe these extra
functionalities.

Signed-off-by: Paul Kocialkowski <paul.kocialkowski@bootlin.com>