media: cedrus: Update TODO with future rework plans
The current TODO list of the cedrus driver is now outdated as most of the points
it mentions were already tackled. In addition it is no longer considered
relevant to wait for a stateless encoder driver to move it out of staging.
The hantro/verisilicon driver was already unstaged without this condition.

However the driver suffers from a bad initial design that showed to be very
limiting. It was also not a very good fit for a video codec driver either.

Since a rework of the driver was already carried out for the purpose of adding
encoding support, update the TODO list with a description of the rework.
This reworked driver will be published soon and transitional commits from the
current driver will be submitted upstreamer after that.

It seems best to wait for the rework to land upstream before unstaging the
driver, since a major rework is best operated on a staging driver instead of a
fully upstream one.

Signed-off-by: Paul Kocialkowski <paul.kocialkowski@bootlin.com>